Renamed view packages to widget packages
