+++ ui/src/main/java/com/mhacks/app/ui/events/EventsFragmentProvider.kt
@@ -1,6 +1,6 @@
 package com.mhacks.app.ui.events
 
-import com.mhacks.app.ui.events.view.EventsFragment
+import com.mhacks.app.ui.events.widget.EventsFragment
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
