--- ui/src/main/java/com/mhacks/app/ui/login/viewpager/LoginViewPagerFragment.kt
@@ -1,25 +1,0 @@
-package com.mhacks.app.ui.login.viewpager
-
-import android.support.v4.app.Fragment
-
-/**
- * Fragment for the main login component.
- * This manages the Views managed by the ViewPager.
- */
-
-class LoginViewPagerFragment: Fragment(){
-//
-//    override fun onCreateView(inflater: LayoutInflater?, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-//        return inflater?.inflate(R.layout.fragment_viewpager_login, container, false)
-//    }
-//
-//    override fun onViewCreated(view: View?, savedInstanceState: Bundle?) {
-//            super.onCreate(savedInstanceState)
-//        login_viewpager.adapter = LoginViewPagerAdapter(context)
-//
-//        pageIndicatorView.setViewPager(login_viewpager)
-//    }
-//    companion object {
-//        val instance get() = LoginViewPagerFragment()
-//    }
-}
