+++ ui/src/main/java/com/mhacks/app/ui/qrscan/QRScanActivity.kt
@@ -32,12 +32,12 @@ import com.google.android.gms.vision.barcode.Barcode
 import com.google.android.gms.vision.barcode.BarcodeDetector
 import com.mhacks.app.extension.viewModelProvider
 import com.mhacks.app.ui.common.BaseActivity
-import com.mhacks.app.ui.qrscan.view.BarcodeGraphic
-import com.mhacks.app.ui.qrscan.view.BarcodeGraphicTracker
-import com.mhacks.app.ui.qrscan.view.BarcodeTrackerFactory
-import com.mhacks.app.ui.qrscan.view.camera.CameraSource
-import com.mhacks.app.ui.qrscan.view.camera.CameraSourcePreview
-import com.mhacks.app.ui.qrscan.view.camera.GraphicOverlay
+import com.mhacks.app.ui.qrscan.widget.BarcodeGraphic
+import com.mhacks.app.ui.qrscan.widget.BarcodeGraphicTracker
+import com.mhacks.app.ui.qrscan.widget.BarcodeTrackerFactory
+import com.mhacks.app.ui.qrscan.widget.camera.CameraSource
+import com.mhacks.app.ui.qrscan.widget.camera.CameraSourcePreview
+import com.mhacks.app.ui.qrscan.widget.camera.GraphicOverlay
 import kotlinx.android.synthetic.main.activity_qr_scan.*
 import org.mhacks.mhacksui.R
 import timber.log.Timber
