+++ ui/src/main/res/layout/collapsible_card_content.xml
@@ -30,7 +30,7 @@
                 android:paddingEnd="16dp"
                 android:paddingStart="16dp"
                 android:paddingTop="8dp"
-                tools:text="HELLO WORLD" />
+                tools:text="Card Title" />
 
             <ImageView
                 android:id="@+id/expand_icon"
