+++ ui/src/main/res/layout/fragment_welcome.xml
@@ -14,6 +14,7 @@
         android:layout_height="match_parent">
 
         <com.github.florent37.arclayout.ArcLayout
+            android:id="@+id/welcome_fragment_arc_layout"
             android:layout_width="match_parent"
             android:layout_height="300dp"
             android:elevation="6dp"
@@ -38,13 +39,19 @@
                     app:defTextSize="@dimen/text_countdown_normaltitles"
                     app:countTextSize="@dimen/text_countdown_timertext"
                     app:marginBtwText= "26dp"
-                    app:barWidth="12dp"
+                    app:barWidth="8dp"
                     app:progressColor="@color/white"
                     app:defTextColor="@color/white"
                     app:countTextColor="@color/white" />
             </FrameLayout>
         </com.github.florent37.arclayout.ArcLayout>
 
+        <android.support.v7.widget.RecyclerView
+            android:id="@+id/welcome_fragment_favorite_events_recycler_view"
+            android:layout_width="match_parent"
+            android:layout_height="0dp"
+            app:layout_constraintTop_toBottomOf="@+id/welcome_fragment_arc_layout"
+            app:layout_constraintBottom_toBottomOf="parent" />
 
     </android.support.constraint.ConstraintLayout>
 </layout>
