+++ ui/src/main/res/values/dimens.xml
@@ -1,13 +1,9 @@
 <resources>
-    <!-- Default screen margins, per the Android Design guidelines. -->
-
-    <!-- Margins -->
-
     <!-- Text Sizes -->
     <dimen name="common_midsize">18sp</dimen>
 
-    <dimen name="text_countdown_normaltitles">24sp</dimen>
-    <dimen name="text_countdown_timertext">36sp</dimen>
+    <dimen name="text_countdown_normaltitles">22sp</dimen>
+    <dimen name="text_countdown_timertext">28sp</dimen>
 
     <!-- Parameters -->
     <dimen name="countdown_progressbar_thickness">10dp</dimen>
